* master: (52 commits)
  [Test] Ignore EtcdEmulatorServiceTest.testGetWait
  English is not added to the supported locales
  Fix disappearing placement group
  Fixing failing tests ClosureBlueprintServiceIT
  [VBV-1192] Call sslSocket.connect in separate threads to unblock xenon
  VBV-1186 - [UI] Create network button disabled and spinning after network create error
  Bug 1841516: Containers Scale Failures -Request failed: maxClauseCount is set to 1024
  Workaround for VBV-1180
  Separate Admiral image into 2, one that is rarely updated (base) and the one that is packaging the binaries
  Revert "Remove placement configuration as part of Closure Description"
  Removed Lightwave-related sources
  VBV-1188 - In compute state the adapterReference should not be the full uri
  Help modal moved to Clarity style.
  [VBV-1177] List most of the stateless services in swagger UI.
  Containers in RETIRED state while running the performance test
  Remove collection trigger buttons
  [VBV-1172] Deleting host doesn`t remove it immediately
  Remove placement configuration as part of Closure Description
  [VCOM-533] Enumeration in azure does not take into account the region id.
  Detach container stats from health check
  ...

+++ test-integration/src/test/java/com/vmware/admiral/host/BaseManagementHostClusterIT.java
@@ -226,7 +226,7 @@ public abstract class BaseManagementHostClusterIT {
     }
 
     protected void tearDownHost(List<ManagementHost> hosts) {
-        hosts.forEach(host -> stopHost(host, false));
+        hosts.forEach(host -> stopHost(host, false, true));
     }
 
 
@@ -242,7 +242,8 @@ public abstract class BaseManagementHostClusterIT {
         }
     }
 
-    protected void stopHost(ManagementHost host, boolean waitWhilePortIsListening) {
+    protected void stopHost(ManagementHost host, boolean waitWhilePortIsListening, boolean
+            deleteStorage) {
 
         if (host == null) {
             return;
@@ -260,12 +261,14 @@ public abstract class BaseManagementHostClusterIT {
                 waitWhilePortIsListening(host, waiter);
                 waiter.await();
             }
-            // Surrounded with try catch because on Windows OS, there is probably
-            // permissions problem and this util cannot delete the sandbox.
-            try {
-                FileUtils.deleteDirectory(new File(host.getStorageSandbox().getPath()));
-            } catch (Throwable t) {
-                logger.log(Level.WARNING, "Error when trying to remove sandbox", t);
+            if (deleteStorage) {
+                // Surrounded with try catch because on Windows OS, there is probably
+                // permissions problem and this util cannot delete the sandbox.
+                try {
+                    FileUtils.deleteDirectory(new File(host.getStorageSandbox().getPath()));
+                } catch (Throwable t) {
+                    logger.log(Level.WARNING, "Error when trying to remove sandbox", t);
+                }
             }
             logger.log(Level.INFO, String.format("Host '%s' stopped", hostname));
 
@@ -283,7 +286,7 @@ public abstract class BaseManagementHostClusterIT {
     protected void stopHostAndRemoveItFromNodeGroup(ManagementHost availableHost,
             ManagementHost hostToStop) {
 
-        stopHost(hostToStop, true);
+        stopHost(hostToStop, true, false);
 
         if (availableHost != null) {
             TestContext ctx = new TestContext(1, Duration.ofMinutes(2));
@@ -476,7 +479,7 @@ public abstract class BaseManagementHostClusterIT {
             logger.log(Level.SEVERE, "Error when trying to start host '" + hostname + "' reason: "
                     + Utils.toString(ex));
             if (retryCount > 0) {
-                stopHost(host, true);
+                stopHost(host, true, false);
                 return startHost(host, sandboxUri, peers, retryCount - 1);
             } else {
                 fail(Utils.toString(ex));
